requirements.txt
openai
python-dotenv
tqdm
anthropic
# probaoy missing somethign 